requirements.txt
asgiref==3.3.4
Django==3.1.7
django-cors-headers==3.7.0
django-treenode==0.17.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Pillow==8.3.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
django-storages==1.11.1
boto3==1.17.99
django-utils-six==2.0
timeago==1.0.15